+++ apollo/src/jsMain/kotlin/io/iohk/atala/prism/apollo/secp256k1/Secp256k1Lib.kt
@@ -4,9 +4,9 @@ import com.ionspin.kotlin.bignum.integer.BigInteger
 import io.iohk.atala.prism.apollo.utils.ECConfig
 import io.iohk.atala.prism.apollo.utils.asByteArray
 import io.iohk.atala.prism.apollo.utils.asUint8Array
-import io.iohk.atala.prism.apollo.utils.decodeHex
 import io.iohk.atala.prism.apollo.utils.external.BN
 import io.iohk.atala.prism.apollo.utils.external.ec
+import io.iohk.atala.prism.apollo.utils.external.secp256k1.SignatureType
 import io.iohk.atala.prism.apollo.utils.external.secp256k1.secp256k1
 import io.iohk.atala.prism.apollo.utils.toHexString
 import org.kotlincrypto.hash.sha2.SHA256
@@ -39,9 +39,7 @@ actual class Secp256k1Lib actual constructor() {
         val derivedPrivKeyString = derivedPrivateKeyBytes.toHexString()
         val privKey = BigInteger.parseString(privKeyString, 16)
         val derivedPrivKey = BigInteger.parseString(derivedPrivKeyString, 16)
-
         val added = (privKey + derivedPrivKey) % ECConfig.n
-
         return added.toByteArray()
     }
 
@@ -53,13 +51,27 @@ actual class Secp256k1Lib actual constructor() {
      * @return The signature of the data in DER format.
      */
     actual fun sign(privateKey: ByteArray, data: ByteArray): ByteArray {
-        // TODO: "Using noble/secp256k1 would be problematic since it requires a configuration so it can use Sync methods to sign, also it doesnt have DER signatures")
         val sha = SHA256().digest(data)
-        val ecInstance = ec("secp256k1")
-        val key = ecInstance.keyFromPrivate(privateKey.asUint8Array())
-        val signature = key.sign(sha.asUint8Array())
-        val derSignature = signature.toDER(enc = "hex").unsafeCast<String>()
-        return derSignature.decodeHex()
+        val signature = secp256k1.sign(
+            sha.asUint8Array(),
+            privateKey.asUint8Array(),
+            {}
+        )
+        return signature.toDERRawBytes().asByteArray()
+    }
+
+    private fun normalise(signatureBytes: ByteArray): SignatureType {
+        val jsSignatureByteArray = signatureBytes.asUint8Array()
+        val signature = try {
+            secp256k1.Signature.fromDER(jsSignatureByteArray)
+        } catch (e: dynamic) {
+            secp256k1.Signature.fromCompact(jsSignatureByteArray)
+        }
+        return if (signature.hasHighS()) {
+            signature.normalizeS()
+        } else {
+            signature
+        }
     }
 
     /**
@@ -75,9 +87,42 @@ actual class Secp256k1Lib actual constructor() {
         signature: ByteArray,
         data: ByteArray
     ): Boolean {
-        val ecjs = ec("secp256k1")
+        val normalised = this.normalise(signature)
         val sha = SHA256().digest(data)
-        return ecjs.verify(sha.toHexString(), signature.toHexString(), publicKey.toHexString(), enc = "hex")
+        if (secp256k1.verify(normalised, sha.asUint8Array(), publicKey.asUint8Array(), {})) {
+            return true
+        }
+        return try {
+            val transcoded = transcodeSignatureToBitcoin(normalised.toCompactRawBytes().asByteArray())
+            secp256k1.verify(transcoded, sha.asUint8Array(), publicKey.asUint8Array(), {})
+        } catch (e: dynamic) {
+            secp256k1.verify(normalised, sha.asUint8Array(), publicKey.asUint8Array(), {})
+        }
+    }
+
+    private fun transcodeSignatureToBitcoin(signature: ByteArray): SignatureType {
+        val rawLen = signature.size / 2
+        val r = signature.copyOfRange(0, rawLen).reversedArray()
+        val s = signature.copyOfRange(rawLen, signature.size).reversedArray()
+        val lenR = r.size
+        val lenS = s.size
+        val derLength = 6 + lenR + lenS
+        val derSignature = ByteArray(derLength)
+        derSignature[0] = 0x30
+        derSignature[1] = (4 + lenR + lenS).toByte()
+        derSignature[2] = 0x02
+        derSignature[3] = lenR.toByte()
+        derSignature.fill(r, 4, lenR + 4)
+        derSignature[4 + lenR] = 0x02
+        derSignature[5 + lenR] = lenS.toByte()
+        derSignature.fill(s, 6 + lenR, lenS + 6 + lenR)
+        return secp256k1.Signature.fromDER(derSignature.asUint8Array())
+    }
+
+    private fun ByteArray.fill(from: ByteArray, start: Int, end: Int) {
+        for (i in start until end) {
+            this[i] = from[i - start]
+        }
     }
 
     /**
@@ -89,12 +134,9 @@ actual class Secp256k1Lib actual constructor() {
     @OptIn(ExperimentalUnsignedTypes::class)
     actual fun uncompressPublicKey(compressed: ByteArray): ByteArray {
         val ecjs = ec("secp256k1")
-
         val decoded = ecjs.curve.decodePoint(compressed.asUint8Array())
-
         val x = ByteArray(decoded.getX().toArray().size) { index -> decoded.getX().toArray()[index].asDynamic() as Byte }
         val y = ByteArray(decoded.getX().toArray().size) { index -> decoded.getY().toArray()[index].asDynamic() as Byte }
-
         val header: Byte = 0x04
         return byteArrayOf(header) + x + y
     }